requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
firebase-admin==6.3.0
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
google-generativeai==0.3.1
scipy==1.11.4
networkx==3.2.1
joblib==1.3.2
gunicorn==21.2.0
